Move api/server to daemon/server
