+++ daemon/server/server_test.go
